Use dynamic clickhouse-cpp library by default
Previously we statically compiled clickhouse-cpp into the extension. But
in general, dynamic libraries are preferred. So build clickhouse-cpp as
a DSO by default, and install it into the package library directory
along with the extension module itself. Compile an `rpath` into the
module so that the OS can find the clickhouse-cpp DSO.

Add the `CH_BUILD` variable that, when set to "static", triggers
statically compiling clickhouse-cpp into the extension module. Update
the CI test to test both configurations.