Always compile `rpath` and require clickhouse-cpp
Update `ci.yaml` to properly clean up the dynamic build before compiling
the static build. This revealed a compile failure on Linux, which
doesn't like `-Wl` without a comma-separated list.

So initially set `SHLIB_LINK` only to the static library or the `-L` and
`-l` values that differ between the dynamic and static builds. Then, set
the rest of the `SHLIB_LINK` list after importing PGXS, and user
`$(pkgconfig)` instead of `pg_config --pkglibdir` to always set the
`rpath`. The static build doesn't need it to find the `clickhouse-cpp`
library, but it's harmless and becoming fairly common among Postgres
extensions.

While at it, add the `$(OBJS)` target to require the vendored
`clickhouse-cpp` submodule, so that it will automatically be initialized
if it hasn't already been. This simplifies building from a fresh clone
of the repository.

In passing, fix a couple of incorrect header levels in the README.