Generate `fdw.c` with version & remove envvar code
Rename `src/fdw.c` to `src/fdw.c.in` and add a `Makefile` target to
generate the former. It replaces `__VERSION__` with the version. Teach
`.gitignore` to ignore the generated file. Inspired by the addition of
`PG_MODULE_MAGIC_EXT` in PostgreSQL 18; see [this post] for details.

Also delete the generated `sql/clickhouse_fdw--2.0.sql` file, and make
sure it and `src/fdw.c` are included in the `clean` target. Teach
`.gitignore` to ignore the generated file, but to keep any files with
two sets of `--` in them, as will be true of upgrade scripts.

While at it, delete code left over from the original experiment using
the envvar extension.

  [this post]: https://justatheory.com/2025/05/pg_module_magic_ext/
    "Postgres Extensions: Use `PG_MODULE_MAGIC_EXT`"