Copy and port old FDW
Fix various warnings from the previous code, mostly unused or shadowed
variables. Also detect the `.c.in` file and append it to a sorted (and
therefor de-duped) list of object files to be sure it gets compiled, as
well. Add `libuuid` to the list of dependencies.

The old clickhouse_fdw uses an old and hacked copy of
yandex/clickhouse-cpp, but here we want to use the latest version of
`clickhouse/clickhouse-cpp`, instead. It varies in a few ways, some of
which are fixed here, others which are commented-out for now with "XXX"
markers.

Also add proper `libcurl` configuration to the `Makefile` and skip
testing on Postgres 18; will update for that version in a separate
commit.