Add SQL to create the extension
And also a few aggregate functions from the previous version of the
extension.

This change causes `installcheck` to actually load the extension, which
failed for a dynamic install, because it wants the symlinked file, and
for static install, because `libcityhash` was missing from `SHLIB_LINK`.
So also change `install-ch-cpp` to use `cp -a` to copy all library files
and preserve symlinks and add `libcityhash` to `SHLIB_LINK`.