Add query tests and fix UUID processing
Copy the `binary.sql` test from the old `clickhouse_fdw` project and fix
issues discovered running it. This version uses `clickhouse_` function
prefixes, rather than `clickhousedb_`. The test is otherwise the same,
but highlighted that UUID support was disabled.

To fix it, replace `UInt128 val` with `auto val` to get the newer
ClickHouse `UUID` type. The interface is otherwise the same, so just
uncomment the rest of the block.

Other changes:

*   Install and start ClickHouse inside the pgxn-tools container for the
    tests to use
*   Replace the `base` test with the `pg_regress --load-extension` option
*   Don't add `luuid` to PG_LDFLAGS on macOS; it seems to be part of the
    OS
*   Update the `FdwModifyPrivateIndex` docs; they were missing an item
*   Add comments for where to add `UPDATE` and `DELETE` support