Add binary queries test
Add `test/sql/binary_queries.sql` from the previous `clickhouse_fdw`.
Exclude the `INSERT` statements, which are not yet supported. Also of
course omit the table and queries related to those statements.

Add variants in expected output for different versions of PostgreSQL, as
well as a `.txt` file that describes the versions of PostgreSQL for
each.

Rename the objects created by `test/sql/binary.sql` and `DROP` them at
the end of the test, so that the tests should be able to run in
parallel.

While at it, always make a static build on macOS (Darwin) unless
`CH_BUILD` is explicitly set. Currently `make installcheck` causes a
segfault with the dynamic library. It would be worthwhile to chase it
down at some point, but this will do for now.

Also borrow the `results` target from pgTAP to simplify updating
expected test output, as well as the `print-*` target to print a `make`
variable's flavor and value.