Properly implement functions and aggregates
Add a new C function, `clickhouse_pushdown`, to be called by functions
and aggregates that fail to be pushed down. The first argument should
describe the thing not pushed down and all other args are optional and
may be of any type.

Use this function in the custom aggregates and functions that should be
pushed down to ClickHouse. `ch_argmax`, `ch_argmin`, and `dictGet` call
it directly.

Change the signatures of `ch_argmax` and `ch_argmin` and the aggregate
that use them, `argMax()` and `argMin()`, to allow two arguments of any
type, as supported by ClickHouse.

Add the `uniqExact` aggregate, which allows any number of arguments of a
variety of types. Coerce it to use the `clickhouse_pushdown` when
pushdown fails by setting the initial value `aggregate uniqExact()` so
that it will be reported as a failed pushdown. Since the aggregate
always returns BIGINT but it uses TEXT for the initial value, add the
`ch_noop_bigint` function which takes a TEXT argument (for the state
being aggregated) and returns a BIGINT. It maps to the new
`clickhouse_noop` function, which simply returns `NULL`.

Since `uniqExact` is a variadic aggregate, remove the code that marked
variadic functions as unsupported and add `deparseArrayList` to rewrite a
variadic array value to a list of values to be passed to ClickHouse.

Change the deparser to match `uniqexact` rather than `uniq_exact`; the
latter seems to have been a custom name used by the previous version of
the FDW, but it seems better to use the exact name that ClickHouse
expects.

Finally, port `test/sql/functions.sql` from the old FDW and update it to
better test the `uniqExact` variadic aggregate. While at it, merge the
files listing expected output files to Postgres versions from files
named for each test to a single file, `test/expected/result_map.txt`.