Fix and test function pushdown failure errors
The `clickhouse_pushdown` function does not work for `argMax` and
`argMin` because they don't return a specific data type, which is
required for the pattern used by `uniqExact` to work. So rewrite them as
PL/pgSQL functions that raise exceptions directly.

While at it, change the error code for all the exceptions to
`fdw_error`, code `HV000`, in both the PL/pgSQL functions and in
`clickhouse_pushdown`. Add a test for these pushdown failures, using a
`DO` statement to keep the output fairly nice and to complain loudly
when an an exception doesn't contain the right code, or when no error is
thrown at all.