Flesh out and test PGXN distribution
Update the `Makefile` target to use `git-archive-all` to create the
archive in order to include `clickhouse-cpp` in the release zip file and
add the `dist-test` target to test a release.

Flesh out `.github/workflows/release.yml` to also use `archive-all` and
also to start Postgres and ClickHouse and test the new release zip file
for every push. Only release on a tag starting with "v".