Map `EXTRACT()` to `date_part()`
Add the `EXTRACT` OIDs to the list of OIDs mapped to `CF_DATE_PART`,
which `deparse.c` converts to ClickHouse equivalents. This allows
`EXTRACT()` to function just like `date_part()`.

Also add the OIDs for extracting values from dates; it previously worked
with `date_part()` for some reason, but did not for `EXTRACT()`. So
include them both for completeness.

Add tests for these patterns. Closes #24.