Add query and status code to error messages
When queries fail, include error detail with the remote query that was
executed. For HTTP queries, limit the query detail to response codes
greater than or equal to 404 and also always include the status code in
error context. Always include the remote SQL for binary queries and
change the error code to `ERRCODE_SQL_ROUTINE_EXCEPTION`.