Set `join_use_nulls`
To increase SQL compatibility, set `join_use_nulls=1` before executing
`SELECT` queries. Discovered via the recommendations for the [TPC-H
queries].

Enabled for the http engine by passing the setting as
a URL query parameter. Enabled for the binary engine by setting it for
each query. It'd be nice to set it for every request via the binary
client, but it does not currently offer an API for that purpose.

These solutions are generalized enough that adding additional settings
in the future will be quite straightforward.

  [TPC-H queries]: https://clickhouse.com/docs/getting-started/example-datasets/tpch#queries