Add Docker image workflow
Add a `Dockerfile` that builds the current branch into a Docker image
for a specific version of Postgres. Then add a `docker-bake.hcl` file
that bakes docker images for each supported Postgres version and for
multiple platforms. Add the `image` target to the `Makefile` to simplify
build the image(s) and optionally pushing to a registry. Finally, add
the `bake.yml` workflow that actually bakes and, when executing for a
tag, pushes the results.

Update `README.md` with up-to-date status badges for the Postgres and
ClickHouse workflows and add one for the Docker image, as well.