Implement binary inserts with BeginInsert flow
Replace the old (and commented-out) code with the new `BeginInsert` /
`InsertData` / `SendInsertBlock` pattern in `clickhouse-cpp`. This
greatly simplifies the code, as the most of the necessary logic for
creating the insert Block object and tracking the state of the insert is
now handled by the Client object.

Remove the old logic for unwrapping `LowCardinality` columns and fix
`column_append` to properly append to those columns. This preserves
traffic over the wire, I'm told. Thanks to @slabko for working out how
to make that work again.

Restore the logic for inserting arrays by creating a column of the
proper type, then use `column_append()` to add the items to it before
appending it to the `ColumnArray`. Thanks to @slabko again for the
incantation to create a column of the proper type for array items.

Copy the test from the old FDW and tweak it to match the other tests.