Add support for ClickHouse Bool
Map the ClickHouse `Bool` data type to the PostgreSQL `BOOLEAN` type.
Internally, the `clickhouse-cpp` library treats booleans as `UInt2`, so
add a custom function, `convert_bool_to_int16`, to cast PostgreSQL
`BOOLEAN` to `INT16` to support inserting into ClickHouse, and tweak
`init_output_convert_state` to use it in this case.

Add tests for boolean inserts and selects to both the binary and HTTP
tests.

With the introduction of proper `Bool` support, remove the `ClickHouse
<UInt8>
type was translated to <INT2>` notice.

While at it:

*   Improve the exception raised by `make_datum() to include the name of
    the unsupported data type. Done to help track down the fact that
    `Bool` wasn't supported.
*   Add the `table_name` parameter to `parse_type()` and update the
    exception it throws to include the table and column for a data type
    that cannot be mapped from a PostgreSQL type to a ClickHouse type.
    Update all the callers to this function as appropriate.