Always `ResetConnection` on error
When using the `clickhouse-cpp` library, always `ResetConnection` on
error and release the insert block. Otherwise, the client can be stuck
in an unusable state that can interfere with future queries. This also
changes the behavior after `EndInsert` fails, changing it to a proper
error rather than a `NOTICE`; it doesn't seem right to let the next
query fail, as the previous implementation had it.