Add JSON support for the HTTP engine
Map it to the `JSON` data type and test table importing, insertion, and
selection. Add a test for both the binary and HTTP engines, and tweak
the ClickHouse configuration to enable the JSON type in 24 and earlier,
although it does not work prior to 24.8.

The Binary engine will successfully import a table with a `JSON` column
but not be able to insert or select that column until `JSON` support is
added to `clickhouse-cpp` (issue ClickHouse/clickhouse-cpp#422; PR
ClickHouse/clickhouse-cpp#450).