Remove `params` function and type
Remove the `params` function that was used to emulate ClickHouse
parametric aggregate arguments. Instead, change `quantile()` and
`quantileExact` to use the PostgreSQL `WITHIN GROUP (ORDER BY x)`
syntax, and deparse them into their ClickHouse parametric equivalents
exactly as `percentile_cont` is translated. This removes a special
casing and funky workaround in favor of translating the differing
syntaxes.

This presumes that Postgres ordered set aggregates are the equivalent of
ClickHouse parametric aggregates, which in the case of `percentile_cont`
and `quantile`, at least, appears correct.