Support the dbname option for http connections
Previously, the HTTP engine never specified a database to connect to,
instead schema-qualifying all tables in generated queries from `dbname`
config for a foreign server. This works well for queries against foreign
tables, but not when using `clickhouse_raw_query()`, which ignored the
`dbname` param.

So add support for an explicit `dbname` when connecting via HTTP by
parsing it into the connection details in `connstring_parse()`, adding
it to the new `data` field in the `ch_http_connection_t` struct when
"connecting" to ClickHouse, and then using that value to set the
`X-ClickHouse-Database` header for all HTTP requests. Change one of the
calls to `clickhouse_raw_query()` in the tests to use this option and
not schema-qualify the query it executes.

While at it, remove the mapping for the `Boolean` type, which doesn't
exist, but is covered by the `Bool` mapping.