Fix missing comma delimiter in `ORDER BY`
Fix a bug where an `ORDER BY` expression with a `NULLS FIRST` followed
by a second column fails to insert a comma between them. This bug is due
to poor indentation thanks to a braceless if/else expression (see also
[Heartblead]). Fix it by emitting an explicit `NULLS LAST`, which
ClickHouse does in fact support.

Add a regression test with an explicit `NULLS FIRST` and a second column
in an `ORDER BY` clause. Update all the expected results for this test
as well as the explicit outputting of `NULLS LAST`.

  [Heartblead]: https://martinfowler.com/articles/testing-culture.html