Pass connection details to driver connectors
Rather than pass each field from `ch_connection_details` to the http and
binary connection functions, define `ch_connection_details` in a new
header, include it in the headers for the drivers, and change their
connection functions to take a `ch_connection_details` directly. This
keeps things tidier and will allow the addition of new fields in the
future without having to change function signatures again.