Add `clickCast` and `accurateCast`
ClickHouse provides a `CAST()` function that operates like PostgreSQL's
cast, but of course the data types are different, and it allows the data
type to be quoted rather than literal. We can't override that function
with a custom function named `cast()`, presumably because there is no
function in Postgres; it's likely recognized by the parser.

So add an alias to ClickHouse's `cast()` function, named `clickCast()`
for clarity. It takes a value and a sting naming a data type and returns
a value of that type. This should cover most of the use cases for the
litany of type-specific [data conversion functions].

While at it, also provide `accurateCast()`, which prevents overflows.

Unfortunately, these functions are limited by the PostgreSQL requirement
that the return type be detectable from the arguments at runtime. That
allows casting from, say, `SMALLINT` to `BIGINT`, or from `TIMESTAMP` to
`TEXT` because they're compatible types. But one cannot currently cast
to incompatible types, such as `clickCast('42'::TEXT, 'Int32')`, because
ClickHouse will return a `Uint32` but Postgres expects `TEXT`, because
it derives the return value type from the first argument.

But at least they're good for compatible types, which is common for
situations such as casting a `UInt8` to `Uint64` to pass to `dictGet()`.

  [data conversion functions]: https://clickhouse.com/docs/sql-reference/functions/type-conversion-functions#cast