Set PG_MODULE_MAGIC_EXT version to full `x.y.z`
The plan is to make `x.y.0` releases that increment `.y` only when there
are SQL changes, and to release `x.y.1+` for library-only changes. This
will allow everyone to get all the library-only bug fixes without having
to run `ALTER EXTENSION UPDATE`.

But while the version listed in the catalog will be `x.y`, it makes
sense to properly version the library with `x.y.z`. So change the value
assigned to `PG_MODULE_MAGIC_EXT.version` to the full release `x.y.z`,
so that users can find the full version, at least on Postgres 18 and
higher.