Use `ColumnDecimal->GetString`
Instead of converting the Int128 value from `clickhouse-cpp`, use the new
`ColumnDecimal->GetString()` function to get the stringification of a
decimal value fetched from ClickHouse. This greatly simplifies the code
and restores the ability to use the `clickhouse-cpp` shared library,
re-enabled in the `Makefile` for all OSes other than Darwin.