Switch to option-style parse for session_settings
The format of the `pg_clickhouse.session_settings` GUC was

    key = val, key = 'val'...

But that doesn't match options passed to `CREATE SERVER` and the like,
which requires space, not an equal sign, between option names and
values:

    key val, key 'val'...

In the future we'll want to merge these options with the GUC, so it's
better if they use the same format.

So add a new parameter to `chfdw_parse_options()`, `bool with_equal`,
and only require an equal sign between the name and value when it't
true. Then change calls to it to not require a comma for the GUC, but to
still require them for the `clickhouse_raw_query()` connection string,
to align with the libpq behavior.

Update the GUC tests and documentation to reflect this change.