Fix the README badges
The ghcr-badge.egpl.dev badge wasn't showing up, perhaps due to some
permissions for the API or something. But it turns out we can get by
without it.

Drop the ghcr-badge.egpl.dev badge in favor of shield.io's GitHub
Release badge, and use it for both the GitHub release and the Docker
Image release. It could get out of sync if we stop releasing the Docker
images or move them, but we'd have to update the badge anyway, so this
seams reasonable.